Moved build file to the location where Unreal expects it.  Rewrote build file to not have to look at the full path for the include.  Added OS X support, but code does not compile in OS X yet
